requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asyncio==3.4.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.19.0
Faker==29.0.0
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-telegram-bot==20.0
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
wheel==0.38.4
yarl==1.11.1
